requirements.txt
gym[mujoco] >= 0.21.0, < 0.24.1
numpy >= 1.20.2
pip install git+https://github.com/Farama-Foundation/d4rl@master#egg=d4rl
jax >= 0.3.13
flax >= 0.5.2
ml_collections >= 0.1.0
tqdm >= 4.60.0
optax >= 0.1.3
absl-py >= 0.12.0
scipy >= 1.6.0
wandb >= 0.12.14
tensorflow-probability >= 0.17.0
moviepy
imageio